refactor db impl
